// src/state/ducks/data/courses/actions/types.ts
export enum CourseDataActionTypes {
    COURSE_FAILURE_ACTION = "course/COURSE_FAILURE_ACTION",
    CREATE_NEW_COURSE = "course/CREATE_NEW_COURSE",
    CREATE_NEW_COURSE_SUCCESS = "course/CREATE_NEW_COURSE_SUCCESS",
    EDIT_COURSE = "course/EDIT_COURSE",
    EDIT_COURSE_SUCCESS = "course/EDIT_COURSE_SUCCESS",
    DELETE_COURSE = "course/DELETE_COURSE",
    DELETE_COURSE_SUCCESS = "course/DELETE_COURSE_SUCCESS",
    GET_COURSES_CURRENT_USER = "course/GET_COURSES_CURRENT_USER",
    GET_COURSES_CURRENT_USER_SUCCESS = "course/GET_COURSES_CURRENT_USER_SUCCESS",
    SET_COURSES_FOR_USER = "course/SET_COURSES_FOR_USER",
}