// src/state/ducks/data/gradeCategories/actions/types.ts
export enum GradeCategoryDataActionTypes {
    CREATE_GRADE_CATEGORY = "gradeCategory/CREATE_GRADE_CATEGORY",
    CREATE_GRADE_CATEGORY_SUCCESS = "gradeCategory/CREATE_GRADE_CATEGORY_SUCCESS",
    GET_GRADE_CATEGORIES_FOR_COURSE = "gradeCategory/GET_GRADE_CATEGORIES_FOR_COURSE",
    GET_GRADE_CATEGORIES_FOR_COURSE_SUCCESS = "gradeCategory/GET_GRADE_CATEGORIES_FOR_COURSE_SUCCESS",
    EDIT_GRADE_CATEGORY = "gradeCategory/EDIT_GRADE_CATEGORY",
    EDIT_GRADE_CATEGORY_SUCCESS = "gradeCategory/EDIT_GRADE_CATEGORY_SUCCESS",
    DELETE_GRADE_CATEGORY = "gradeCategory/DELETE_GRADE_CATEGORY",
    DELETE_GRADE_CATEGORY_SUCCESS = "gradeCategory/DELETE_GRADE_CATEGORY_SUCCESS",
    GET_GRADE_CATEGORIES_FOR_CURRENT_USER = "gradeCategory/GET_GRADE_CATEGORIES_FOR_CURRENT_USER",
    GET_GRADE_CATEGORIES_FOR_CURRENT_USER_SUCCESS = "gradeCategory/GET_GRADE_CATEGORIES_FOR_CURRENT_USER_SUCCESS",
    SET_GRADE_CATEGORIES_FOR_USER = "gradeCategory/SET_GRADE_CATEGORIES_FOR_USER",

    CREATE_GRADE = "grades/CREATE_GRADE",
    CREATE_GRADE_SUCCESS = "grades/CREATE_GRADE_SUCCESS",
    EDIT_GRADE = "grades/EDIT_GRADE",
    EDIT_GRADE_SUCCESS = "grades/EDIT_GRADE_SUCCESS",
    DELETE_GRADE = "grades/DELETE_GRADE",
    DELETE_GRADE_SUCCESS = "grades/DELETE_GRADE_SUCCESS",

    GRADE_CATEGORY_FAILURE_ACTION = "gradeCategory/GRADE_CATEGORY_FAILURE_ACTION",
}